Rename types.
